Use robot to test TCP URL support.
