+++ tcp/src/test/resources/org/kaazing/net/impl/tcp/echo.then.closed.rpt
@@ -1,0 +1,6 @@
+accept tcp://localhost:61234
+accepted
+connected
+read  "Hello, world"
+write "Hello, world"
+closed
